Init commit for client and notes about boiler plate
This project uses a react hot boilerplate from here:
https://github.com/gaearon/react-hot-boilerplate

It's a open source projet and does nothing more that provide us
with the neccessary "blank slate" or boilerplate for our project.
It is nothing more than a starting point for our project.

We imagine that this is in accordance with the rules and guidelines
for HACK UCSC.